// packages/plugin-timeclock-ui/src/types.ts
export interface ITimelog {
  _id: string;
  timelog: Date;
  user: IUser;
  deviceSerialNo?: string;
  deviceName?: string;
}